+++ packages/airless-google-cloud-bigquery/airless/google/cloud/bigquery/operator/__init__.py
@@ -1,7 +1,3 @@
-from .bigquery import (GcsQueryToBigqueryOperator, PubsubToBigqueryOperator, FileToBigqueryOperator)
+from .bigquery import GcsQueryToBigqueryOperator, PubsubToBigqueryOperator
 
-__all__ = [
-    'GcsQueryToBigqueryOperator',
-    'PubsubToBigqueryOperator',
-    'FileToBigqueryOperator'
-]
+__all__ = ['GcsQueryToBigqueryOperator', 'PubsubToBigqueryOperator']